+++ src/NexusLabs.Eve/EveStreamFollowMode.cs
@@ -1,0 +1,7 @@
+﻿namespace NexusLabs.Eve;
+
+internal enum EveStreamFollowMode
+{
+    ActiveTurnResponse,
+    SessionStream,
+}